TSP: use longs consistently for distance to avoid overflow (fix regression during City to Location refactor)

+++ optaplanner-examples/src/main/java/org/optaplanner/examples/tsp/domain/location/AirLocation.java
@@ -33,10 +33,10 @@ public class AirLocation extends Location {
     }
 
     @Override
-    public int getDistance(Location location) {
+    public long getDistance(Location location) {
         double distance = getAirDistanceDouble(location);
         // Multiplied by 1000 to avoid floating point arithmetic rounding errors
-        return (int) (distance * 1000.0 + 0.5);
+        return (long) (distance * 1000.0 + 0.5);
     }
 
 }
