+++ optaplanner-examples/src/main/java/org/optaplanner/examples/tsp/domain/location/RoadLocation.java
@@ -47,13 +47,13 @@ public class RoadLocation extends Location {
     }
 
     @Override
-    public int getDistance(Location location) {
+    public long getDistance(Location location) {
         if (this == location) {
-            return 0;
+            return 0L;
         }
         double distance = travelDistanceMap.get((RoadLocation) location);
         // Multiplied by 1000 to avoid floating point arithmetic rounding errors
-        return (int) (distance * 1000.0 + 0.5);
+        return (long) (distance * 1000.0 + 0.5);
     }
 
 }
